fix-#51: SearchResult 관련 컴포넌트 PlaylistComponent에서 분리, ColorCategory 관련 컴포넌트 PrimaryPlayListTabScreen class로 이동
