+++ presentation/src/main/java/com/pyc/presentation/playlist/view/ui/primaryplaylist/PrimaryPlayListTabScreen.kt
@@ -1,4 +1,4 @@
-package com.pyc.presentation.playlist.view.ui.components
+package com.pyc.presentation.playlist.view.ui.primaryplaylist
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
@@ -16,16 +16,62 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.pyc.presentation.model.ColorModel
 import com.pyc.presentation.R
+import com.pyc.presentation.model.AudioPlaylistItemModel
+import com.pyc.presentation.model.ColorModel
+import com.pyc.presentation.playlist.view.ui.components.PlaylistScreenComponent
 
 
 @Composable
-internal fun ColorListRow(
+internal fun PrimaryPlayListScreen(
+    colorList: List<ColorModel>,
+    playlist: List<AudioPlaylistItemModel>,
+    playlistId: Int,
+    itemClick: (Int, AudioPlaylistItemModel) -> Unit,
+    itemLongClick: (Int) -> Unit,
+    moreIconClick: (AudioPlaylistItemModel) -> Unit,
+    onAddAudio: (Int) -> Unit,
+    onEditPlaylist: (Int) -> Unit,
+    onSearchAudioInList: (String) -> Unit
+) {
+    //임시
+    val selectedColorId: Int? = 1
+
+    PlaylistScreenComponent(
+        topBarSlot = {
+            ColorCategoryListRow(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(
+                        horizontal = dimensionResource(id = R.dimen.item_padding_horizontal),
+                        vertical = dimensionResource(id = R.dimen.item_padding_vertical)
+                    ),
+                colorList,
+                selectedColorId,
+                onAllSelect = {},
+                onColorSelect = {})
+        },
+        playlist = playlist,
+        playlistId = playlistId,
+        itemClick = itemClick,
+        itemLongClick = itemLongClick,
+        moreIconClick = moreIconClick,
+        onAddAudio = onAddAudio,
+        onEditPlaylist = onEditPlaylist,
+        onSearchAudioInList = onSearchAudioInList
+
+
+    )
+}
+
+
+@Composable
+internal fun ColorCategoryListRow(
     modifier: Modifier,
     colorList: List<ColorModel>,
     selectedColorId: Int?,
@@ -59,7 +105,7 @@ internal fun ColorListRow(
                 .weight(0.8f)
         ) {
             itemsIndexed(colorList) { index, color ->
-                ColorItemColumn(selectedColorId, item = color)
+                ColorCategoryItemColumn(selectedColorId, item = color)
             }
         }
 
@@ -68,7 +114,7 @@ internal fun ColorListRow(
 
 
 @Composable
-internal fun ColorItemColumn(selectedColorId: Int?, item: ColorModel) {
+internal fun ColorCategoryItemColumn(selectedColorId: Int?, item: ColorModel) {
     Column(
         modifier = Modifier
             .wrapContentSize(Alignment.Center)
@@ -92,4 +138,7 @@ internal fun ColorItemColumn(selectedColorId: Int?, item: ColorModel) {
         }
     }
 }
+
+
+
